backend/requirements.txt
SQLAlchemy==2.0.36
fastapi==0.115.6
sec-api==1.0.25
psycopg2==2.9.10
alembic==1.14.0
pydantic==2.10.3
python-dotenv == 1.0.1
uvicorn==0.34.0